+++ CHANGELOG.md
@@ -19,6 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the `JSON` response class
 - Added the `custom` body translate strategy
 - Made `method` a keyword-only parameter in `path`
+- Added the `extract_path` utility
+- Added the `view build` command
+- Added `App.template`
 - **Breaking Change:** The `body` parameter in `Response` is now required
 
 ## [1.0.0-alpha8] - 2024-1-21